ADD: staging environemtn

+++ deploy/ci.sh
@@ -3,18 +3,32 @@
 USERNAME=${CREDENTIALS%:*}
 PASSWORD=${CREDENTIALS#*:}
 
-sed -e "s;%BUILD_NUMBER%;${BUILD_NUMBER};g" ./deploy/kube/rcs/myapp.yaml > temp.yaml
+if [ "$PASSENGER_APP_ENV" = "production" ]; then
 
-LAST_SUCCESSFUL_BUILD=$(curl http://${USERNAME}:${PASSWORD}@52.38.170.255:8080/job/myapp/lastSuccessfulBuild/buildNumber)
+  sed -e "s;%BUILD_NUMBER%;${BUILD_NUMBER};g" ./deploy/kube/rcs/myapp-production.yaml > temp.yaml
+  LAST_SUCCESSFUL_BUILD=$(curl http://${USERNAME}:${PASSWORD}@52.38.170.255:8080/job/myapp-production/lastSuccessfulBuild/buildNumber)
+  /usr/bin/kubectl rolling-update myapp-v${LAST_SUCCESSFUL_BUILD} -f temp.yaml --namespace production
+  if [ $? -eq 0 ]
+  then
+    echo "Deployment successful"
+  else
+    echo "Deployment stopped."
+    exit 1
+  fi
+  rm temp.yaml
 
-/usr/bin/kubectl rolling-update myapp-v${LAST_SUCCESSFUL_BUILD} -f temp.yaml --namespace production
+elif [ "$PASSENGER_APP_ENV" = "staging" ]; then
 
-if [ $? -eq 0 ]
-then
-  echo "Deployment successful"
-else
-  echo "Deployment stopped."
-  exit 1
+  sed -e "s;%BUILD_NUMBER%;${BUILD_NUMBER};g" ./deploy/kube/rcs/myapp-staging.yaml > temp.yaml
+  LAST_SUCCESSFUL_BUILD=$(curl http://${USERNAME}:${PASSWORD}@52.38.170.255:8080/job/myapp-staging/lastSuccessfulBuild/buildNumber)
+  /usr/bin/kubectl rolling-update myapp-v${LAST_SUCCESSFUL_BUILD} -f temp.yaml --namespace staging
+  if [ $? -eq 0 ]
+  then
+    echo "Deployment successful"
+  else
+    echo "Deployment stopped."
+    exit 1
+  fi
+rm temp.yaml
 fi
 
-rm temp.yaml
